init: Reservation Controller, service, repository 생성자 주입
Fixes: #53

+++ src/main/java/com/commit/campus/entity/Reservation.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reservationId;  // 예약 ID
+    private int reservationId;  // 예약 ID
 
     @ManyToOne
     @JoinColumn(name = "camp_facs_id")
